Migrate to qdb http client
